Create README.md
Every day in the morning we need to complete some set of predefined tasks that remain
constant. We can list the generic tasks as below:
1. Ring the alarm on time
2. Make coffee
3. Heat the water to a suitable temperature
4. Pack your bag (Keep only appropriate books for the day)
5. Cook breakfast and lunch
6. Iron the clothes

ABOUT OF ROBOT:-
Consider the Robot as your personal assistant. You will give the Robot a name by which you will
call it. You also need to define the functions performed by the Robot. These functions will be
the same as above i.e. the tasks which it will perform for you in the morning. Feel free to
add/modify the above-listed tasks as per your choice.